+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The tax calculations for NFTs have been corrected, they are no longer "same day"
 - Added static typing. ([#230](https://github.com/BittyTax/BittyTax/issues/230))
 - Added support for Python 3.12.
 - Binance parser: added "Transaction Fee" operation for statements. ([#314](https://github.com/BittyTax/BittyTax/issues/314))
+- Etherscan parser: added new export formats. ([#313](https://github.com/BittyTax/BittyTax/issues/313))
 ### Changed
 - Accounting tool: use openpyxl instead of xlrd for reading .xlsx files. ([#260](https://github.com/BittyTax/BittyTax/issues/260))
 - Conversion tool: use openpyxl instead of xlrd for reading .xlsx files. ([#260](https://github.com/BittyTax/BittyTax/issues/260))